+++ test/cmdlineTests.sh
@@ -673,6 +673,18 @@ SOLTMPDIR=$(mktemp -d)
 )
 rm -r "$SOLTMPDIR"
 
+printTask "Testing EVM Assembly JSON import/export..."
+SOLTMPDIR=$(mktemp -d)
+(
+    cd "$SOLTMPDIR"
+    if ! "$REPO_ROOT/scripts/ASTImportTest.sh" evm-assembly
+    then
+        rm -r "$SOLTMPDIR"
+        fail
+    fi
+)
+rm -r "$SOLTMPDIR"
+
 printTask "Testing AST import/export..."
 SOLTMPDIR=$(mktemp -d)
 (